// vanila_redux_todo-list/src/enteties/character/model/actionTypes/characterActionTypes.ts
export enum CharacterActionTypes {
  SET_IS_LOADING = "CHARACTER/SET_IS_LOADING",
  SET_ERROR = "CHARACTER/SET_ERROR",
  SET_CHARACTERS = "CHARACTER/SET_CHARACTER",
  SET_SINGLE_CHARACTER = "CHARACTER/SET_SINGLE_CHARACTER",
  CLEAR_ALL_CHARACTER_META = "CHARACTER/CLEAR_ALL_CHARACTER_META",
  SET_CHARACTER_SEARCH_QUERY = "CHARACTER/SET_CHARACTER_SEARCH_QUERY",
  SET_CHARACTER_PAGE = "CHARACTER/SET_CHARACTER_PAGE",
  SET_CHARACTER_PAGE_BACK = "CHARACTER/SET_CHARACTER_PAGE_BACK",
  SET_CHARACTER_PAGE_FORWARD = "CHARACTER/SET_CHARACTER_PAGE_FORWARD",
  SET_CHARACTER_INFO = "CHARACTER/SET_CHARACTER_INFO",
}